Core compiles right now
Former-commit-id: 8a6acec00e222490561469c2f54aac9dacb1002c
Former-commit-id: 63fac0dfb9691276a73a51d89cd596fdffb33715

+++ Source/FlounderEngine/src/worlds/Worlds.cpp
@@ -78,12 +78,12 @@ namespace Flounder
 
 	float Worlds::GetSunriseFactor() const
 	{
-		return Maths::Clamp(-(sin(2.0f * PI * GetDayFactor()) - 1.0f) / 2.0f, 0.0f, 1.0f);
+		return Maths::Clamp(static_cast<float>(-(sin(2.0f * PI * GetDayFactor()) - 1.0f)) / 2.0f, 0.0f, 1.0f);
 	}
 
 	float Worlds::GetShadowFactor() const
 	{
-		return Maths::Clamp(1.7f * sin(2.0f * PI * GetDayFactor()), 0.0f, 1.0f);
+		return Maths::Clamp(static_cast<float>(1.7f * sin(2.0f * PI * GetDayFactor())), 0.0f, 1.0f);
 	}
 
 	float Worlds::GetSunHeight() const
